requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.0
webdriver-manager>=4.0.1
playwright>=1.40.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.2

# Database
sqlalchemy>=2.0.23

# Web Server
flask>=3.0.0
gunicorn>=21.0.0

# Utilities
python-dotenv>=1.0.0
fake-useragent>=1.4.0
tenacity>=8.2.3
tqdm>=4.66.1

# Anti-bot detection
undetected-chromedriver>=3.5.4

# Rate limiting
ratelimit>=2.2.1

# Proxy support
httpx>=0.25.2

# Web Interface
flask>=3.0.0
